+++ nutzboot-demo/nutzboot-demo-feign/nutzboot-demo-feign-service/src/main/java/io/nutz/demo/feign/MainLauncher.java
@@ -1,0 +1,24 @@
+package io.nutz.demo.feign;
+
+import org.nutz.boot.NbApp;
+import org.nutz.dao.Dao;
+import org.nutz.ioc.loader.annotation.Inject;
+import org.nutz.ioc.loader.annotation.IocBean;
+
+import io.nutz.demo.feign.bean.User;
+
+@IocBean(create = "init")
+public class MainLauncher {
+
+    @Inject
+    protected Dao dao;
+
+    public void init() {
+        dao.create(User.class, true);
+    }
+
+    public static void main(String[] args) throws Exception {
+        new NbApp().setPrintProcDoc(true).run();
+    }
+
+}