+++ nutzboot-demo/nutzboot-demo-cloud/nutzboot-demo-cloud-service/src/main/java/io/nutz/demo/feign/CloudServiceLauncher.java
@@ -4,15 +4,24 @@ import org.nutz.boot.NbApp;
 import org.nutz.dao.Dao;
 import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
+import org.nutz.lang.util.NutMap;
+import org.nutz.mvc.annotation.At;
+import org.nutz.mvc.annotation.Ok;
 
 import io.nutz.demo.feign.bean.User;
 
 @IocBean(create = "init")
-public class MainLauncher {
+public class CloudServiceLauncher {
 
     @Inject
     protected Dao dao;
 
+    @Ok("raw")
+    @At("/Status")
+    public NutMap status() {
+        return new NutMap();
+    }
+    
     public void init() {
         dao.create(User.class, true);
     }