+++ nutzboot-starter/nutzboot-starter-feign/src/main/java/org/nutz/boot/starter/feign/FeignStarter.java
@@ -1,7 +1,11 @@
 package org.nutz.boot.starter.feign;
 
 import java.lang.reflect.Field;
+import java.util.List;
 
+import javax.inject.Provider;
+
+import org.nutz.boot.AppContext;
 import org.nutz.boot.annotation.PropDoc;
 import org.nutz.boot.starter.feign.annotation.FeignInject;
 import org.nutz.ioc.Ioc;
@@ -9,9 +13,22 @@ import org.nutz.ioc.IocEventListener;
 import org.nutz.ioc.impl.PropertiesProxy;
 import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
+import org.nutz.lang.Mirror;
 import org.nutz.lang.Strings;
-import org.nutz.log.Log;
-import org.nutz.log.Logs;
+
+import com.netflix.client.config.DefaultClientConfigImpl;
+import com.netflix.discovery.EurekaClient;
+import com.netflix.loadbalancer.IRule;
+import com.netflix.loadbalancer.LoadBalancerBuilder;
+import com.netflix.loadbalancer.RandomRule;
+import com.netflix.loadbalancer.ServerList;
+import com.netflix.loadbalancer.ServerListFilter;
+import com.netflix.loadbalancer.ServerListUpdater;
+import com.netflix.loadbalancer.ZoneAffinityServerListFilter;
+import com.netflix.loadbalancer.ZoneAwareLoadBalancer;
+import com.netflix.niws.loadbalancer.DiscoveryEnabledNIWSServerList;
+import com.netflix.niws.loadbalancer.DiscoveryEnabledServer;
+import com.netflix.niws.loadbalancer.EurekaNotificationServerListUpdater;
 
 import feign.Client;
 import feign.Feign;
@@ -22,7 +39,6 @@ import feign.codec.Encoder;
 import feign.gson.GsonDecoder;
 import feign.gson.GsonEncoder;
 import feign.httpclient.ApacheHttpClient;
-import feign.hystrix.FallbackFactory;
 import feign.hystrix.HystrixFeign;
 import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonEncoder;
@@ -30,13 +46,13 @@ import feign.jaxb.JAXBContextFactory;
 import feign.jaxb.JAXBDecoder;
 import feign.jaxb.JAXBEncoder;
 import feign.okhttp.OkHttpClient;
+import feign.ribbon.LBClient;
+import feign.ribbon.LBClientFactory;
 import feign.ribbon.RibbonClient;
 import feign.slf4j.Slf4jLogger;
 
 @IocBean
 public class FeignStarter implements IocEventListener {
-    
-    private static final Log log = Logs.get();
 
     protected static String PRE = "feign.";
 
@@ -63,6 +79,9 @@ public class FeignStarter implements IocEventListener {
 
     @Inject("refer:$ioc")
     protected Ioc ioc;
+    
+    @Inject
+    protected AppContext appContext;
 
     @Inject
     protected PropertiesProxy conf;
@@ -70,6 +89,7 @@ public class FeignStarter implements IocEventListener {
     @SuppressWarnings({"unchecked", "rawtypes"})
     public Object afterBorn(Object obj, String beanName) {
         try {
+            Mirror mirror = Mirror.me(obj);
             for (Field field : obj.getClass().getDeclaredFields()) {
                 FeignInject fc = field.getAnnotation(FeignInject.class);
                 if (fc == null)
@@ -79,7 +99,7 @@ public class FeignStarter implements IocEventListener {
                 Client client = getClient(fc, field);
                 Logger.Level level = Level.valueOf(conf.get(PROP_LOGLEVEL, "BASIC").toUpperCase());
                 String url = Strings.sBlank(conf.get(PROP_URL), "http://127.0.0.1:8080");
-                Class<?> apiType = field.getType();
+                Class apiType = field.getType();
                 Logger logger = new Slf4jLogger(apiType);
 
                 boolean useHystrix = "true".equals(Strings.sBlank(fc.useHystrix(), conf.get(PROP_HYSTRIX_ENABLE)));
@@ -92,21 +112,13 @@ public class FeignStarter implements IocEventListener {
                     builder.client(client);
                 builder.logger(logger);
                 builder.logLevel(level);
-                Object t = useHystrix ? ((HystrixFeign.Builder) builder).target(apiType, url, new FallbackFactory() {
-                    public Object create(Throwable cause) {
-                        try {
-                            if (Strings.isBlank(fc.fallback()))
-                                return ioc.getByType(apiType);
-                            return ioc.get(apiType, fc.fallback());
-                        }
-                        catch (Exception e) {
-                            log.debugf("fallback to ioc bean type=[%s], but not found or error", apiType.getName());
-                            return null;
-                        }
-                    }
-                }) : builder.target(apiType, url);
-                field.setAccessible(true);
-                field.set(obj, t);
+                Object t = null;
+                if (useHystrix) {
+                    t = ((HystrixFeign.Builder) builder).target(apiType, url, getIocBean(apiType, fc.fallback()));
+                } else {
+                    t = builder.target(apiType, url);
+                }
+                mirror.setValue(obj, field.getName(), t);
             }
         }
         catch (Throwable e) {
@@ -137,7 +149,7 @@ public class FeignStarter implements IocEventListener {
             return new GsonDecoder();
         case "jaxb":
             JAXBContextFactory jaxbFactory = new JAXBContextFactory.Builder().withMarshallerJAXBEncoding("UTF-8")
-                                                                             .withMarshallerSchemaLocation(Strings.sBlank(fc.schema(), conf.get(PROP_SCHEMA)))
+                                                                             .withMarshallerSchemaLocation(getSchemaString(fc.schema()))
                                                                              .build();
             return new JAXBDecoder(jaxbFactory);
         default:
@@ -158,7 +170,9 @@ public class FeignStarter implements IocEventListener {
         case "gson":
             return new GsonEncoder();
         case "jaxb":
-            JAXBContextFactory jaxbFactory = new JAXBContextFactory.Builder().withMarshallerJAXBEncoding("UTF-8").withMarshallerSchemaLocation(fc.schema()).build();
+            JAXBContextFactory jaxbFactory = new JAXBContextFactory.Builder().withMarshallerJAXBEncoding("UTF-8")
+                                                                             .withMarshallerSchemaLocation(getSchemaString(fc.schema()))
+                                                                             .build();
             return new JAXBEncoder(jaxbFactory);
         default:
             break;
@@ -167,7 +181,7 @@ public class FeignStarter implements IocEventListener {
     }
 
     protected Client getClient(FeignInject fc, Field field) {
-        switch (Strings.sBlank(fc.client(), conf.get(PROP_CLIENT, "jdk"))) {
+        switch (getClientString(fc)) {
         case "jdk":
             // nop
             break;
@@ -176,10 +190,53 @@ public class FeignStarter implements IocEventListener {
         case "httpclient":
             return new ApacheHttpClient();
         case "ribbon":
-            return RibbonClient.create();
+            return RibbonClient.builder().lbClientFactory(new LBClientFactory() {
+                public LBClient create(String clientName) {
+                    return getLoadBalancer(clientName);
+                }
+            }).build();
         default:
             break;
         }
         return null;
+    }
+
+    public String getClientString(FeignInject fc) {
+        return Strings.sBlank(fc.client(), conf.get(PROP_CLIENT, "jdk"));
+    }
+
+    public <T> T getIocBean(Class<T> apiType, String fallbackName) {
+        if (!Strings.isBlank(fallbackName))
+            return ioc.get(apiType, fallbackName);
+        List<T> list = appContext.getBeans(apiType);
+        if (list.size() > 0)
+            return list.get(0);
+        return null;
+    }
+
+    public String getSchemaString(String schema) {
+        return Strings.sBlank(schema, conf.get(PROP_SCHEMA));
+    }
+    
+    public LBClient getLoadBalancer(String name) {
+        Provider<EurekaClient> provider = new Provider<EurekaClient>() {
+            public EurekaClient get() {
+                return ioc.get(EurekaClient.class, "eurekaClient");
+            }
+        };
+        IRule rule = new RandomRule();
+        DefaultClientConfigImpl clientConfig = DefaultClientConfigImpl.getClientConfigWithDefaultValues(name);
+        ServerList<DiscoveryEnabledServer> list = new DiscoveryEnabledNIWSServerList(name, provider);
+        ServerListFilter<DiscoveryEnabledServer> filter = new ZoneAffinityServerListFilter<DiscoveryEnabledServer>(clientConfig);
+        ServerListUpdater updater = new EurekaNotificationServerListUpdater(provider);
+        ZoneAwareLoadBalancer<DiscoveryEnabledServer> lb = LoadBalancerBuilder.<DiscoveryEnabledServer>newBuilder()
+                .withDynamicServerList(list)
+                .withRule(rule)
+                .withServerListFilter(filter)
+                .withServerListUpdater(updater)
+                .withClientConfig(clientConfig)
+                .buildDynamicServerListLoadBalancerWithUpdater();
+        lb.enableAndInitLearnNewServersFeature();
+        return LBClient.create(lb, clientConfig);
     }
 }