+++ nutzboot-contrib/nutzboot-starter-ureport/src/main/java/org/nutz/boot/starter/ureport/UreportServletStarter.java
@@ -15,11 +15,11 @@ import com.bstek.ureport.console.UReportServlet;
 public class UreportServletStarter extends UReportServlet implements WebServletFace {
 
     @PropDoc(value = "定义UReport2中提供的默认基于文件系统的报表存储目录", defaultValue = "")
-    public static final String UREPORT_FILESTOREDIR = "ureport.fileStoreDir";
+    public static final String PROP_FILESTOREDIR = "ureport.fileStoreDir";
 
 
     @PropDoc(value = "是否启用自定义报表存储器", defaultValue = "")
-    public static final String UREPORT_DISABLEFILEPROVIDER= "ureport.disableFileProvider";
+    public static final String PROP_DISABLEFILEPROVIDER= "ureport.disableFileProvider";
 
 
     public String getName() {