+++ ChangeLog.md
@@ -2,9 +2,11 @@
 
 ## dev
 
-变更:
+* 变更:
 	* add: [caffeine](https://github.com/ben-manes/caffeine) by [幸福的旁边](https://github.com/happyday517)
+	* add: [hystrix](https://github.com/Netflix/Hystrix) 支持@HystrixCommand和/hystrix.stream
 	* change: feign默认不设置encoder/decoder,上一个版本默认jackson,并添加更多配置项.
+	* change: feign添加feign-hystrix依赖,强化与hystrix的集成
 
 ## 2.0.1 "刚好遇见你"