+++ ChangeLog.md
@@ -6,9 +6,12 @@
 	* add: [caffeine](https://github.com/ben-manes/caffeine) by [幸福的旁边](https://github.com/happyday517)
 	* add: [hystrix](https://github.com/Netflix/Hystrix) 支持@HystrixCommand和/hystrix.stream
 	* add: [hystrix-dashboard](https://github.com/Netflix/Hystrix)
+	* add: [eureka-server](https://github.com/Netflix/eureka) 服务治理
+	* add: [eureka-client](https://github.com/Netflix/eureka) 
 	* add: [xxl-job-executor](https://github.com/xuxueli/xxl-job) 
+	* add: feign完成feign-ribbon-eureka集成,实现负载均衡和服务发现
+	* add: feign添加feign-hystrix依赖,强化与hystrix的集成
 	* change: feign默认不设置encoder/decoder,上一个版本默认jackson,并添加更多配置项.
-	* change: feign添加feign-hystrix依赖,强化与hystrix的集成
 	* change: 重构项目结构,用户项目不再需要依赖nutzboot-starter